grid from string tested
